refactor: page 컴포넌트 pages 폴더로 이동
